Another small bit in ComputationGraph / MultiLayer that's factorizable
