+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/impl/graph/SparkComputationGraph.java
@@ -18,27 +18,22 @@
 
 package org.deeplearning4j.spark.impl.graph;
 
-import lombok.NonNull;
 import org.apache.spark.SparkContext;
 import org.apache.spark.api.java.JavaDoubleRDD;
 import org.apache.spark.api.java.JavaPairRDD;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.rdd.RDD;
-import org.deeplearning4j.api.storage.StatsStorageRouter;
-import org.deeplearning4j.api.storage.StatsStorageRouterProvider;
-import org.deeplearning4j.api.storage.listener.RoutingIterationListener;
 import org.deeplearning4j.nn.conf.ComputationGraphConfiguration;
 import org.deeplearning4j.nn.graph.ComputationGraph;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
-import org.deeplearning4j.optimize.api.IterationListener;
 import org.deeplearning4j.spark.api.TrainingMaster;
 import org.deeplearning4j.spark.api.stats.SparkTrainingStats;
+import org.deeplearning4j.spark.impl.SparkListenable;
 import org.deeplearning4j.spark.impl.common.reduce.IntDoubleReduceFunction;
 import org.deeplearning4j.spark.impl.graph.dataset.DataSetToMultiDataSetFn;
 import org.deeplearning4j.spark.impl.graph.dataset.PairDataSetToMultiDataSetFn;
 import org.deeplearning4j.spark.impl.graph.scoring.*;
-import org.deeplearning4j.spark.impl.listeners.VanillaStatsStorageRouterProvider;
 import org.deeplearning4j.spark.util.SparkUtils;
 import org.deeplearning4j.util.ModelSerializer;
 import org.nd4j.linalg.api.ndarray.INDArray;
@@ -51,14 +46,10 @@ import org.nd4j.linalg.heartbeat.reports.Environment;
 import org.nd4j.linalg.heartbeat.reports.Event;
 import org.nd4j.linalg.heartbeat.reports.Task;
 import org.nd4j.linalg.heartbeat.utils.EnvironmentUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import scala.Tuple2;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.Serializable;
-import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
@@ -66,19 +57,15 @@ import java.util.concurrent.atomic.AtomicInteger;
  *
  * @author Alex Black
  */
-public class SparkComputationGraph implements Serializable {
-    private static final Logger log = LoggerFactory.getLogger(SparkComputationGraph.class);
+public class SparkComputationGraph extends SparkListenable {
 
     public static final int DEFAULT_EVAL_SCORE_BATCH_SIZE = 64;
     private transient JavaSparkContext sc;
-    private TrainingMaster trainingMaster;
     private ComputationGraphConfiguration conf;
     private ComputationGraph network;
     private double lastScore;
 
     private transient AtomicInteger iterationsCount = new AtomicInteger(0);
-
-    private List<IterationListener> listeners = new ArrayList<>();
 
     /**
      * Instantiate a ComputationGraph instance with the given context and network.
@@ -274,74 +261,6 @@ public class SparkComputationGraph implements Serializable {
     @Deprecated
     public ComputationGraph fitMultiDataSet(String path, int minPartitions) {
         return fitMultiDataSet(path);
-    }
-
-    /**
-     * This method allows you to specify IterationListeners for this model.
-     *
-     * @param listeners Iteration listeners
-     */
-    public void setListeners(@NonNull Collection<IterationListener> listeners) {
-        this.listeners.clear();
-        this.listeners.addAll(listeners);
-        if (trainingMaster != null) trainingMaster.setListeners(this.listeners);
-    }
-
-    /**
-     * This method allows you to specify IterationListeners for this model.
-     * Note that for listeners like StatsListener (that have state that will be sent somewhere), consider instead
-     * using {@link #setListeners(StatsStorageRouter, Collection)}
-     *
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(@NonNull IterationListener... listeners) {
-        setListeners(Arrays.asList(listeners));
-    }
-
-    /**
-     * Set the listeners, along with a StatsStorageRouter that the results will be shuffled to (in the case of any listeners
-     * that implement the {@link RoutingIterationListener} interface)
-     *
-     * @param statsStorage Stats storage router to place the results into
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(StatsStorageRouter statsStorage, IterationListener... listeners) {
-        setListeners(statsStorage, Arrays.asList(listeners));
-    }
-
-    /**
-     * Set the listeners, along with a StatsStorageRouter that the results will be shuffled to (in the case of any listeners
-     * that implement the {@link RoutingIterationListener} interface)
-     *
-     * @param statsStorage Stats storage router to place the results into
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(StatsStorageRouter statsStorage, Collection<? extends IterationListener> listeners) {
-        //Check if we have any RoutingIterationListener instances that need a StatsStorage implementation...
-        StatsStorageRouterProvider routerProvider = null;
-        if(listeners != null ){
-            for(IterationListener l : listeners){
-                if(l instanceof RoutingIterationListener){
-                    RoutingIterationListener rl = (RoutingIterationListener)l;
-                    if(rl.getStorageRouter() == null){
-                        log.warn("RoutingIterationListener provided without providing any StatsStorage instance. Iterator may not function without one. Listener: {}", l);
-                    } else if(!(rl.getStorageRouter() instanceof Serializable)){
-                        //Spark would throw a (probably cryptic) serialization exception later anyway...
-                        throw new IllegalStateException("RoutingIterationListener provided with non-serializable storage router");
-                    }
-
-                    //Need to give workers a router provider...
-                    if(routerProvider == null){
-                        routerProvider = new VanillaStatsStorageRouterProvider();
-                    }
-                }
-            }
-        }
-        this.listeners.clear();
-        if(listeners != null) {
-            this.listeners.addAll(listeners);
-            if (trainingMaster != null) trainingMaster.setListeners(statsStorage, this.listeners);
-        }
     }
 
     /**
