+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/impl/multilayer/SparkDl4jMultiLayer.java
@@ -39,6 +39,7 @@ import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
 import org.deeplearning4j.optimize.api.IterationListener;
 import org.deeplearning4j.spark.api.TrainingMaster;
 import org.deeplearning4j.spark.api.stats.SparkTrainingStats;
+import org.deeplearning4j.spark.impl.SparkListenable;
 import org.deeplearning4j.spark.impl.common.reduce.IntDoubleReduceFunction;
 import org.deeplearning4j.spark.impl.listeners.VanillaStatsStorageRouterProvider;
 import org.deeplearning4j.spark.impl.multilayer.evaluation.*;
@@ -71,16 +72,13 @@ import java.util.*;
  * @author Adam Gibson, Alex Black
  */
 @Slf4j
-public class SparkDl4jMultiLayer implements Serializable {
+public class SparkDl4jMultiLayer extends SparkListenable {
     public static final int DEFAULT_EVAL_SCORE_BATCH_SIZE = 64;
     public static final int DEFAULT_ROC_THRESHOLD_STEPS = 32;
     private transient JavaSparkContext sc;
-    private TrainingMaster trainingMaster;
     private MultiLayerConfiguration conf;
     private MultiLayerNetwork network;
     private double lastScore;
-
-    private List<IterationListener> listeners = new ArrayList<>();
 
     /**
      * Instantiate a multi layer spark instance
@@ -286,77 +284,6 @@ public class SparkDl4jMultiLayer implements Serializable {
     public MultiLayerNetwork fitContinuousLabeledPoint(JavaRDD<LabeledPoint> rdd) {
         return fit(MLLibUtil.fromContinuousLabeledPoint(sc, rdd));
     }
-
-    /**
-     * This method allows you to specify IterationListeners for this model.
-     * Note that for listeners like StatsListener (that have state that will be sent somewhere), consider instead
-     * using {@link #setListeners(StatsStorageRouter, Collection)}
-     *
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(@NonNull Collection<IterationListener> listeners) {
-        setListeners(null, listeners);
-    }
-
-    /**
-     * This method allows you to specify IterationListeners for this model.
-     * Note that for listeners like StatsListener (that have state that will be sent somewhere), consider instead
-     * using {@link #setListeners(StatsStorageRouter, Collection)}
-     *
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(@NonNull IterationListener... listeners){
-        setListeners(Arrays.asList(listeners));
-    }
-
-    /**
-     * Set the listeners, along with a StatsStorageRouter that the results will be shuffled to (in the case of any listeners
-     * that implement the {@link RoutingIterationListener} interface)
-     *
-     * @param statsStorage Stats storage router to place the results into
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(StatsStorageRouter statsStorage, IterationListener... listeners) {
-        setListeners(statsStorage, Arrays.asList(listeners));
-    }
-
-    /**
-     * Set the listeners, along with a StatsStorageRouter that the results will be shuffled to (in the case of any listeners
-     * that implement the {@link RoutingIterationListener} interface)
-     *
-     * @param statsStorage Stats storage router to place the results into
-     * @param listeners    Listeners to set
-     */
-    public void setListeners(StatsStorageRouter statsStorage, Collection<? extends IterationListener> listeners) {
-        //Check if we have any RoutingIterationListener instances that need a StatsStorage implementation...
-        StatsStorageRouterProvider routerProvider = null;
-        if(listeners != null) {
-            for(IterationListener l : listeners) {
-                if(l instanceof RoutingIterationListener) {
-                    RoutingIterationListener rl = (RoutingIterationListener)l;
-                    if(rl.getStorageRouter() == null){
-                        log.warn("RoutingIterationListener provided without providing any StatsStorage instance. Iterator may not function without one. Listener: {}", l);
-                    } else if(!(rl.getStorageRouter() instanceof Serializable)) {
-                        //Spark would throw a (probably cryptic) serialization exception later anyway...
-                        throw new IllegalStateException("RoutingIterationListener provided with non-serializable storage router");
-                    }
-
-                    //Need to give workers a router provider...
-                    if(routerProvider == null){
-                        routerProvider = new VanillaStatsStorageRouterProvider();
-                    }
-                }
-            }
-        }
-
-        this.listeners.clear();
-        if(listeners != null) {
-            this.listeners.addAll(listeners);
-            if (trainingMaster != null) trainingMaster.setListeners(statsStorage, this.listeners);
-        }
-    }
-
-
 
     /**
      * Gets the last (average) minibatch score from calling fit. This is the average score across all executors for the
